Extract logic on getting fields information for MS SQL Server
